test: consolidate duplicated test helpers into tests/helpers.py
content_of (4 copies), make_context (2 copies), and the attribute-based
make_hx (shared by the headers and attributes unit tests) were copy-pasted
across files. Move them to tests/helpers.py and import from there. The two
genuinely specialized make_hx factories (get_triggers wiring in
test_triggers_unit, WidgetFormHx in test_form_unit) stay local.
